fix: Batches not polling during upload if on a different tab
